company_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
prometheus-client==0.19.0
aio-pika==9.3.1
